Fixed bugs on equipment selection, shop and crash on login.

+++ app/src/main/java/com/example/rpggame/activity/ClanLeaderActivity.java
@@ -167,12 +167,38 @@ public class ClanLeaderActivity extends AppCompatActivity {
                 });
     }
     private void showDeleteDialog() {
-        new AlertDialog.Builder(this)
-                .setTitle("Delete Clan")
-                .setMessage("Are you sure you want to delete this clan? This action cannot be undone.")
-                .setPositiveButton("Delete", (dialog, which) -> deleteClan())
-                .setNegativeButton("Cancel", null)
-                .show();
+
+        db.collection("missions")
+                .whereEqualTo("idSaveza", clanId)
+                .get()
+                .addOnSuccessListener(querySnapshot ->{
+                    boolean activeMission = false;
+
+                    for (DocumentSnapshot doc : querySnapshot.getDocuments()) {
+                        String status = doc.getString("status");
+                        if ("AKTIVNA".equals(status)) {   // or whatever value means it's ongoing
+                            activeMission = true;
+                            break;
+                        }
+                    }
+
+                    if (activeMission) {
+                        new AlertDialog.Builder(this)
+                                .setTitle("You cannot leave now!")
+                                .setMessage("You cannot leave clan while clan mission is active.")
+                                .show();
+                    } else{
+                        new AlertDialog.Builder(this)
+                                .setTitle("Delete Clan")
+                                .setMessage("Are you sure you want to delete this clan? This action cannot be undone.")
+                                .setPositiveButton("Delete", (dialog, which) -> deleteClan())
+                                .setNegativeButton("Cancel", null)
+                                .show();
+                    }
+                }).addOnFailureListener(e -> {
+                    Toast.makeText(this, "Failed to check mission: " + e.getMessage(), Toast.LENGTH_SHORT).show();
+                });
+
     }
 
     private void deleteClan() {
